+++ src/LargeTextField.tsx
@@ -2,6 +2,11 @@ import { styled, TextField, TextFieldProps } from '@mui/material';
 import { FC } from 'react';
 import { rem } from './helpers';
 
+/**
+ * Large size variant of TextField
+ *
+ * @deprecated Use <TextField size="large"/> instead
+ */
 export const LargeTextField: FC<TextFieldProps> = styled(TextField)<TextFieldProps>({
   '& .MuiOutlinedInput-input': {
     fontSize: rem(16),