+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 - 2023-03-13
+
+- Added `size="large"` variant to `<TextField/>`
+- Deprecated `<LargeTextField/>` component
+
 ## 1.0.0 - 2023-03-12
 
 - Added `createCompactTheme()` function